Now builds cart in portions as becomes possible, plus a few minor fixes:
(1) Now builds the empty-basket notice and spreadcart icon upon page load, but only builds the "filled" portion of the cart upon getting a valid apiBasketId, which happens at least by the time an item is added to the cart.

(2) Fixed failure to update quantity upon adding to cart, which problem I introduced with the last pull request.

(3) Removed shadow from shopping cart edge. I think frills really ought to be added by the user, without having to be removed to be replaced with something else the user prefers, as I'm having to do.

(4) Defaulted demo index.html to spreadCartIcon -- forgot to do that with the last checkin.