Largely revised to decouple from SpreadShop JS and use Spreadshirt APIs:
(1) Now reads the basket from the Spreadshirt API, no longer from local storage, except for the apiBasketId. The basket is read into a new class SpreadCartBasket, whose data structure mirrors much of the API XML. The node.js proxy now supports a 'read' action to read the cart from the SpreadShop API. The PHP proxy will need to add this action.

(2) Upon receiving a failure to read the basket from the API, assumes that the basket is now invalid and empty. In particular, this happens upon returning to the web site after checkout.

(3) Changed proxy 'operation' parameter to 'action' parameter to match your good choice of terms in the PHP proxy. Done in both spreadCart.js and nodejs/cart.js. You'll need to update your PHP proxy.

(4) Upon adding an item to the basket, timeouts run in series reading the basket from the API until either a maximum number of tries or we get a basket with a different quantity of items. updateMills and updateTries configure this.

(5) If the new 'decouple' configuration variable is true, the timeout series will run on every click made to a page that contains the spreadshop config variable spread_shop_config. There is no dependency on SpreadShop CSS IDs. This mode also allows spreadCart to synchronize with the SpreadShop basket.

(6) No longer dependent on the event onSpreadShopLoaded. Instead runs a timeout series at page load waiting for the necessary components to appear.

(7) The basket structure is built in its entirety at page load and its contents assigned when visible, instead of building as basket data becomes ready.

(8) When an item is deleted, the item display is immediately removed and the basket is reread from the API to update totals. This is particularly important for updating shipping cost.

(9) If the user ever deletes anything from the shopping cart while on a page that has SpreadShop, the page reloads when the user closes the cart, in order to force SpreadShop to use the modified shopping cart. We'll need to do the same when changing item quantities.

(10) Changed font of #totalQuantity to sans-serif, which is more readable when small.

(11) Replaced hardcoded "font-size:60%" in shopping cart footer notes with class="miniBasketNote", added class to CSS.

(12) Spaced the cart total row a bit from the shipping row.

(13) Changed #checkoutLink to #miniBasketCheckoutLink to reduce chances of conflict with SpreadShop or hosting site.

(14) url-encoded the 'emptyBasketUrl' checkout parameter

(15) Made basketItemName, basketItemPrice, and miniBasketLabel (totals lines) all the same font size (110%). Made miniBasketNote font size 75%.

(16) Added UTC timestamp and node.js server error log.

(17) Added "-offline" option to node.js server to have all requests return public/offline.html.